--- webpage/homepage.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-
-    </head>
-    <body>
-        <div id="header">
-            <a href="/predonkulator">Predonkulate</a>
-        </div>
-        <div id="footer">
-            
-        </div>
-    </body>
-</html>
